EVG-15758: move repobuilder from Curator to Barque
